+++ application/controllers/Peminjaman.php
@@ -9,31 +9,54 @@ class Peminjaman extends Render_Controller
     public function index()
     {
         // Page Settings
+        $brg = $this->db->get_where('barang', ['stok' > 0])->result_array();
+        $key = $this->input->post('key');
         $this->navigation = ['Dashboard'];
         $this->plugins = [];
         $this->data['data'] = $this->model->get();
-        $this->data['barang'] = $this->db->get_where('barang', ['stok' > 0])->result_array();
+        $this->data['barang'] = $brg;
         if ($this->session->userdata('id_level') == 1) {
             $this->content = 'peminjaman-teknisi';
         } elseif ($this->session->userdata('id_level') == 2) {
-            $this->content = 'peminjaman-petugas-gudang';
+            $this->content = 'peminjaman-kepala-gudang';
         } elseif ($this->session->userdata('id_level') == 3) {
             $this->content = 'peminjaman-kepala-gudang';
         } elseif ($this->session->userdata('id_level') == 4) {
             $this->content = 'peminjaman-manajer';
         }
+        if ($key) {
+            $this->content = 'peminjaman-barang';
+            $this->data['barang'] = $this->db->like('nama', $key)->get('barang')->result_array();
+        }
         $this->title = $this->content;
+        $this->render();
+    }
+    public function detail($id)
+    {
+        $this->title = 'Detail Peminjaman';
+        $this->navigation = ['Dashboard'];
+        $this->plugins = [];
+        $this->data['barang'] = $this->db->get_where('barang', ['kode_barang' => $id])->row_array();
+        $this->content = 'barang-detail';
         $this->render();
     }
     public function ubah($id)
     {
-        $this->title = 'Ubah Peminjaman';
-        $this->navigation = ['Dashboard'];
-        $this->plugins = [];
         $this->data['dataID'] = $this->model->getById($id);
         $this->data['barang'] = $this->db->get_where('barang', ['stok' > 0])->result_array();
-        $this->content = 'peminjaman-ubah';
-        $this->render();
+        if ($this->data['dataID']['status'] == 'accepted') {
+            echo "<script>alert('Data tidak dapat diubah karna sudah disetujui')</script>";
+            redirect('peminjaman', 'refresh');
+        } else if ($this->data['dataID']['status'] == 'rejected') {
+            echo "<script>alert('Data tidak dapat diubah karna sudah ditolak')</script>";
+            redirect('peminjaman', 'refresh');
+        } else {
+            $this->title = 'Ubah Peminjaman';
+            $this->navigation = ['Dashboard'];
+            $this->plugins = [];
+            $this->content = 'peminjaman-ubah';
+            $this->render();
+        }
     }
     public function accept_gudang($id)
     {
@@ -148,10 +171,19 @@ class Peminjaman extends Render_Controller
     }
     public function hapus($id)
     {
-        $query = $this->model->delete($id);
-        if ($query) {
-            echo "<script>alert('Berhasil Dihapus')</script>";
+        $data = $this->db->get_where('peminjaman', ['id_peminjaman' => $id])->row_array();
+        if ($data['status'] == 'accepted') {
+            echo "<script>alert('Data tidak dapat dihapus karna sudah disetujui')</script>";
             redirect('peminjaman', 'refresh');
+        } else if ($data['status'] == 'rejected') {
+            echo "<script>alert('Data tidak dapat dihapus karna sudah ditolak')</script>";
+            redirect('peminjaman', 'refresh');
+        } else {
+            $query = $this->model->delete($id);
+            if ($query) {
+                echo "<script>alert('Berhasil Dihapus')</script>";
+                redirect('peminjaman', 'refresh');
+            }
         }
     }
